Cleaning unecessary files/folders (part1)
In the existence of documents, the examples of input files can be
deleted. (a complete example will be availabe in the doc representing
the same example)

For the time_series files, a new function is added to the code to load
those data. the data files itself are moved to a new directory as
follow:

ramp/example