modifications on core.py
1. Adding the docs to the main APIs
2. some changes on the code including:

changes on UseCase
===================
add_user
	improvement in error handling by capturing non-User object as
user

__init__
	improving initializing users

changes on User
===============
__repr__
	adding the __repr__ method

save
	in the previous version, the save method was giving vague error
in case no appliance was given (error related to pd.concat of None).
instead a clearer error message is raise in this case

changes on Appliance
====================
__init__
	removing the p_series argument for identifying if the power is
series or not. instead isinstance function is used to identify the type
of the input and catch the possible errrors on the data type or the size
of the data

__repr__
	adding the __repr__ method