Adding the paper example input_files to ramp.example
changes:
1. example files

The three files of the example related to the ramp publication are added
to ramp.example package.

2. download_example function

As the package will be installed through the distribution files,
probably it wont be easy for user to find the example scripts.
To ease the issue, a function is added to ramp.exampl.example to copy
the files to a user given directory.

3. docs

the api doc for the new changes are updated