Function download_example does not work (#66)
Issue opened originally by @LorenzoRinaldi.

Description of the issue:
"download_example" function (ramp\example\examples.py) and it seems there is an issue with the "shower_p.csv" file, which i guess it is supposed to be present in the folder "site-packages\ramp\example" but it is not. the issue is related to the fact that ".csv" files are not included in the package_data argument in setup.py.

Changes:

1. setup.py:
	adding "*.csv" to the packge_data

2. ramp/core/example/examples.py --> function "download_example":
	adding all the csv files of the example to the list of
downloadable files

Test:

The new change is tested locally by creating the source dist files and
installing using pip install -e .