Parameter to switch on/off continuous operation within duty_cycles
When this parameter is set to False (default) te behavior should
be the same as previously, i.e. continuous duty cyles during the
entire switch on event (can contain several duty cycles).

Only when set to True would the switch on event be limited to the
duration of the duty cycle in order to better model productive use
of appliances (mill, welder)