User empty object print ValueError
issue:
When a user object is empty cannot be printed due as the __repr__ method
uses the save() method to print the properties of the users, which raise
Exception in case no appliances are added to the user.

solution:
in case save() methods returns an exception, the function prints the
user_name, and user_num with a message to mention no appliances are
assigned to the user object yet.