The changes introduce a new feature that allows to implement load changes over years.
A new input "num_years" has been added. The multi-year feature is based on the same structure as for the "cases". At the moment I have removed the "cases" option, but I can add it back in depending on how would you like to proceed. The multi-year feature projects random increases in the number of users and appliances over the years. Additionally, new appliances can be defined according to the excel file. I have added a template file. Initially I tried to implement in such a way that the new appliances are defined inside an excel file. I have attached a template along with the python code to call the new function and plot the load. Future iterations could include not only increases but also decreases in load as well as changes in the consumption patterns.